requirement.txt
Django==4.0.5
sqlparse==0.4.2
tzdata==2022.1
djangorestframework
markdown
django-filter

django-role-permissions
django-cors-headers
drf-yasg
djoser